# Dnd.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Dnd.API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/Campaign": {
      "get": {
        "tags": [
          "Campaign"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Campaign"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICampaign"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ICampaign"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ICampaign"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Campaign"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICampaign"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ICampaign"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ICampaign"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Campaign/{id}": {
      "get": {
        "tags": [
          "Campaign"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Campaign"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Campaign/{id}/sessions": {
      "get": {
        "tags": [
          "Campaign"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Campaign"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICampaignSession"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ICampaignSession"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ICampaignSession"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Campaign/{id}/sessions/{sessionId}": {
      "delete": {
        "tags": [
          "Campaign"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Campaign"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICampaignSession"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ICampaignSession"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ICampaignSession"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Characters": {
      "get": {
        "tags": [
          "Character"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Character"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Characters/{id}": {
      "get": {
        "tags": [
          "Character"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/ping": {
      "get": {
        "tags": [
          "Public"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/secretping": {
      "get": {
        "tags": [
          "Public"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/classes": {
      "get": {
        "tags": [
          "Public"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/roll": {
      "post": {
        "tags": [
          "Roll"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RollRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RollRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RollResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/roll/dice": {
      "post": {
        "tags": [
          "Roll"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiceRollRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DiceRollRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DiceRollRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DiceRollResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiceRollResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiceRollResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/roll/dice/simulate": {
      "post": {
        "tags": [
          "Roll"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiceSimulationRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DiceSimulationRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DiceSimulationRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DiceSimulationResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiceSimulationResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiceSimulationResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/{username}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CharacterRequestDto": {
        "type": "object",
        "properties": {
          "abilityScores": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "skillModifiers": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "proficiencies": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "classes": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "ac": {
            "type": "integer",
            "format": "int32"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CharacterResponseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "stats": {
            "$ref": "#/components/schemas/CharacterStats"
          },
          "proficiencies": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "ac": {
            "type": "integer",
            "format": "int32"
          },
          "classes": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CharacterStats": {
        "type": "object",
        "properties": {
          "abilityScores": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "abilityModifiers": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "skillModifiers": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "proficiencies": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ClassResponseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "spellcastingAbility": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DescriptiveStatistics": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "mean": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "variance": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "standardDeviation": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "skewness": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "kurtosis": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "maximum": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "minimum": {
            "type": "number",
            "format": "double",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DiceRollRequestDto": {
        "type": "object",
        "properties": {
          "d4": {
            "type": "integer",
            "format": "int32"
          },
          "d6": {
            "type": "integer",
            "format": "int32"
          },
          "d8": {
            "type": "integer",
            "format": "int32"
          },
          "d10": {
            "type": "integer",
            "format": "int32"
          },
          "d12": {
            "type": "integer",
            "format": "int32"
          },
          "d20": {
            "type": "integer",
            "format": "int32"
          },
          "d100": {
            "type": "integer",
            "format": "int32"
          },
          "modifier": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DiceRollResponseDto": {
        "type": "object",
        "properties": {
          "d4": {
            "type": "integer",
            "format": "int32"
          },
          "d6": {
            "type": "integer",
            "format": "int32"
          },
          "d8": {
            "type": "integer",
            "format": "int32"
          },
          "d10": {
            "type": "integer",
            "format": "int32"
          },
          "d12": {
            "type": "integer",
            "format": "int32"
          },
          "d20": {
            "type": "integer",
            "format": "int32"
          },
          "d100": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "modifier": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DiceSet": {
        "type": "object",
        "properties": {
          "d4": {
            "type": "integer",
            "format": "int32"
          },
          "d6": {
            "type": "integer",
            "format": "int32"
          },
          "d8": {
            "type": "integer",
            "format": "int32"
          },
          "d10": {
            "type": "integer",
            "format": "int32"
          },
          "d12": {
            "type": "integer",
            "format": "int32"
          },
          "d20": {
            "type": "integer",
            "format": "int32"
          },
          "d100": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DiceSimulationRequestDto": {
        "type": "object",
        "properties": {
          "d4": {
            "type": "integer",
            "format": "int32"
          },
          "d6": {
            "type": "integer",
            "format": "int32"
          },
          "d8": {
            "type": "integer",
            "format": "int32"
          },
          "d10": {
            "type": "integer",
            "format": "int32"
          },
          "d12": {
            "type": "integer",
            "format": "int32"
          },
          "d20": {
            "type": "integer",
            "format": "int32"
          },
          "d100": {
            "type": "integer",
            "format": "int32"
          },
          "modifier": {
            "type": "integer",
            "format": "int32"
          },
          "trials": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DiceSimulationResponseDto": {
        "type": "object",
        "properties": {
          "diceRolled": {
            "$ref": "#/components/schemas/IDiceSet"
          },
          "modifier": {
            "type": "integer",
            "format": "int32"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ISimResult"
            },
            "nullable": true
          },
          "trials": {
            "type": "integer",
            "format": "int32"
          },
          "stats": {
            "$ref": "#/components/schemas/DescriptiveStatistics"
          }
        },
        "additionalProperties": false
      },
      "ICampaign": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IDomainUser"
            },
            "nullable": true
          },
          "owner": {
            "$ref": "#/components/schemas/IDomainUser"
          },
          "dungeonMaster": {
            "$ref": "#/components/schemas/IDomainUser"
          }
        },
        "additionalProperties": false
      },
      "ICampaignSession": {
        "type": "object",
        "properties": {
          "campaign": {
            "$ref": "#/components/schemas/ICampaign"
          },
          "sessionDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ICharacter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "ac": {
            "type": "integer",
            "format": "int32"
          },
          "classes": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "stats": {
            "$ref": "#/components/schemas/CharacterStats"
          },
          "proficiencyModifier": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "IDiceSet": {
        "type": "object",
        "properties": {
          "d4": {
            "type": "integer",
            "format": "int32"
          },
          "d6": {
            "type": "integer",
            "format": "int32"
          },
          "d8": {
            "type": "integer",
            "format": "int32"
          },
          "d10": {
            "type": "integer",
            "format": "int32"
          },
          "d12": {
            "type": "integer",
            "format": "int32"
          },
          "d20": {
            "type": "integer",
            "format": "int32"
          },
          "d100": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "IDomainUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "username": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "characters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ICharacter"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ISimResult": {
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "frequency": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "RollRequestDto": {
        "type": "object",
        "properties": {
          "rollType": {
            "type": "string",
            "nullable": true
          },
          "ability": {
            "type": "string",
            "nullable": true
          },
          "characterId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "modifier": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "diceRolled": {
            "$ref": "#/components/schemas/DiceSet"
          },
          "classUsed": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RollResponseDto": {
        "type": "object",
        "properties": {
          "rollType": {
            "type": "string",
            "nullable": true
          },
          "rollValue": {
            "type": "integer",
            "format": "int32"
          },
          "characterId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "characterName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}